move performance tests to test/performance
