+++ tests/performance/suites/basic/config_env_test.go
@@ -1,4 +1,4 @@
-package performance
+package basic
 
 import (
 	"bufio"
