+++ tests/performance/suites/basic/dependencies_test.go
@@ -1,4 +1,4 @@
-package performance
+package basic
 
 import (
 	"fmt"
