+++ tests/performance/suites/basic/helpers_test.go
@@ -1,4 +1,4 @@
-package performance
+package basic
 
 import (
 	"context"
