+++ tests/performance/suites/basic/k6_helpers_test.go
@@ -1,4 +1,4 @@
-package performance
+package basic
 
 import (
 	"fmt"
