+++ tests/performance/suites/basic/performance_suite_test.go
@@ -1,4 +1,4 @@
-package performance
+package basic
 
 import (
 	"fmt"
